+++ src/DynamoCore/BuiltInAndOperators/BuiltIn.Migrations.xml
@@ -10,7 +10,13 @@
   </priorNameHint>
   <priorNameHint>
     <oldName>ImportFromCSV</oldName>
-    <newName>DSCore.IO.Data.ImportCSV</newName>
+    <newName>DSOffice.Data.ImportCSV</newName>
+    <additionalAttributes>
+      <attribute>
+        <name>assembly</name>
+        <value>DSOffice.dll</value>
+      </attribute>
+    </additionalAttributes>
   </priorNameHint>
   <priorNameHint>
     <oldName>AllFalse</oldName>
