+++ src/Libraries/CoreNodes/DSCoreNodes.Migrations.xml
@@ -22,11 +22,23 @@
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.File.ExportToCSV</oldName>
-    <newName>DSCore.IO.File.ExportToCSV</newName>
+    <newName>DSOffice.Data.ExportCSV</newName>
+    <additionalAttributes>
+      <attribute>
+        <name>assembly</name>
+        <value>DSOffice.dll</value>
+      </attribute>
+    </additionalAttributes>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.CSV.WriteToFile</oldName>
-    <newName>DSCore.IO.Data.ExportCSV</newName>
+    <newName>DSOffice.Data.ExportCSV</newName>
+    <additionalAttributes>
+      <attribute>
+        <name>assembly</name>
+        <value>DSOffice.dll</value>
+      </attribute>
+    </additionalAttributes>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.Directory.Contents</oldName>
